+++ src/validator/Session.ts
@@ -1,15 +1,9 @@
 import { NextFunction, Request, Response } from 'express';
-import * as createError from 'http-errors';
 import * as Joi from 'joi';
-import * as _ from 'lodash';
-
-interface IValidations {
-  schema?: Joi.ObjectSchema;
-  data: object;
-}
+import { validator } from './util';
 
 const schemasValidation = {
-  post: Joi.object({
+  POST: Joi.object({
     session: Joi.string().required(),
     user_id: Joi.string().required(),
     device: Joi.string().optional(),
@@ -17,36 +11,13 @@ const schemasValidation = {
   })
 };
 
-const validation = (options: IValidations): void => {
-  const { schema, data } = options;
-
-  if (!schema) return;
-
-  const { error } = schema.validate(data);
-
-  if (error) {
-    const errorMess = _.get(error, 'details[0].message', 'Bad Request!');
-    throw new createError.BadRequest(errorMess);
-  }
-
-  return;
-};
-
 export default async (req: Request, res: Response, next: NextFunction) => {
   try {
     const method = req.method;
     const body = req.body;
-    // const id = req.params;
-    // const params = req.params;
-    // const query = req.query;
+    const schemaValidator: Joi.ObjectSchema = schemasValidation[method];
 
-    const validations = {
-      POST: () => validation({ schema: schemasValidation.post, data: body })
-    };
-
-    if (validations[method]) {
-      validations[method].call();
-    }
+    await validator(schemaValidator, body, method);
 
     next();
   } catch (error) {
